requirements.txt
Bootstrap_Flask==2.4.1
Flask_CKEditor==1.0.0
Flask_Login==0.6.3
Flask_WTF==1.2.1
WTForms==3.1.2
Werkzeug==3.0.4
Flask==3.0.3
flask_sqlalchemy==3.1.1
SQLAlchemy==2.0.35  
gunicorn==23.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
email-validator==2.2.0
openai==1.76.0
pandas==2.2.3
openpyxl==3.1.5
markdown==3.7